feat: Add crews handler with profile, public, get, create, update, and delete endpoints
