+++ src/database/database-do.ts
@@ -175,99 +175,17 @@ export class DatabaseDO extends DurableObject<Env> {
 				}
 			}
 
-			// Crew endpoints
-			if (endpoint === '/crews/profile') {
-				const address = url.searchParams.get('address');
-				if (!address) {
-					return createApiResponse('Missing address parameter', 400);
+			// pass off to handler
+			if (endpoint.startsWith('/crews')) {
+				const handler = getHandler(endpoint);
+				if (handler) {
+					return handler({
+						orm: this.orm,
+						env: this.env,
+						request,
+						url,
+					});
 				}
-
-				// Get the session token from Authorization header
-				const authHeader = request.headers.get('Authorization');
-				if (!authHeader) {
-					return createApiResponse('Missing authorization header', 401);
-				}
-
-				// Extract token from Bearer format
-				const token = authHeader.replace('Bearer ', '');
-
-				// Verify the token matches the requested address
-				const tokenAddress = await validateSessionToken(this.env, token);
-				if (!tokenAddress.success || tokenAddress.address !== address) {
-					return createApiResponse('Unauthorized access', 403);
-				}
-
-				const crews = await getCrewsByProfile(this.orm, address);
-				return createApiResponse({
-					message: 'Successfully retrieved profile crews',
-					data: { crews },
-				});
-			}
-
-			if (endpoint === '/crews/public') {
-				const crews = await getPublicCrews(this.orm);
-				return createApiResponse({
-					message: 'Successfully retrieved public crews',
-					data: { crews },
-				});
-			}
-
-			if (endpoint === '/crews/get') {
-				const crewId = url.searchParams.get('id');
-				if (!crewId) {
-					return createApiResponse('Missing id parameter', 400);
-				}
-				const crew = await getCrew(this.orm, parseInt(crewId));
-				return createApiResponse({
-					message: 'Successfully retrieved crew',
-					data: { crew },
-				});
-			}
-
-			if (endpoint === '/crews/create') {
-				if (request.method !== 'POST') {
-					return createApiResponse('Method not allowed', 405);
-				}
-				const crewData = (await request.json()) as Omit<UserCrewsTable, 'id' | 'created_at' | 'updated_at'>;
-				if (!crewData.profile_id || !crewData.crew_name) {
-					return createApiResponse('Missing required fields: profile_id, crew_name', 400);
-				}
-				const crew = await createCrew(this.orm, crewData);
-				return createApiResponse({
-					message: 'Successfully created crew',
-					data: { crew },
-				});
-			}
-
-			if (endpoint === '/crews/update') {
-				if (request.method !== 'PUT') {
-					return createApiResponse('Method not allowed', 405);
-				}
-				const crewId = url.searchParams.get('id');
-				if (!crewId) {
-					return createApiResponse('Missing id parameter', 400);
-				}
-				const updates = (await request.json()) as Partial<Omit<UserCrewsTable, 'id' | 'created_at' | 'updated_at' | 'profile_id'>>;
-				const result = await updateCrew(this.orm, parseInt(crewId), updates);
-				return createApiResponse({
-					message: 'Successfully updated crew',
-					data: { result },
-				});
-			}
-
-			if (endpoint === '/crews/delete') {
-				if (request.method !== 'DELETE') {
-					return createApiResponse('Method not allowed', 405);
-				}
-				const crewId = url.searchParams.get('id');
-				if (!crewId) {
-					return createApiResponse('Missing id parameter', 400);
-				}
-				const result = await deleteCrew(this.orm, parseInt(crewId));
-				return createApiResponse({
-					message: 'Successfully deleted crew',
-					data: { result },
-				});
 			}
 
 			// Agent endpoints
